refactor: remove useUser
