--- apps/website/src/app/login/layout.tsx
@@ -1,21 +1,0 @@
-"use client"
-
-import { useUser } from "@/lib/client"
-
-import { redirect } from "next/navigation"
-import { toast } from "sonner"
-
-export default function Layout({
-  children,
-}: Readonly<{ children: React.ReactNode }>) {
-  const user = useUser()
-  if (user) {
-    toast("You are already logged in")
-    redirect("/")
-  }
-  return (
-    <div className="flex h-full w-full items-center justify-center">
-      {children}
-    </div>
-  )
-}
