+++ apps/website/src/lib/client.ts
@@ -1,11 +1,4 @@
-import { useSession } from "next-auth/react"
 import { useTheme as _useTheme } from "next-themes"
-
-// TODO: remove this
-export function useUser() {
-  const { data: session } = useSession()
-  return session?.user
-}
 
 export function useTheme() {
   const { theme: _theme, setTheme: _setTheme } = _useTheme()
